requirements.txt
flask
flask-cors
requests
pandas
geopandas
scikit-learn
faker
paho-mqtt
Shapely
